+++ crypto/CMakeLists.txt
@@ -323,9 +323,10 @@ if(ENABLE_DILITHIUM)
   set(
     DILITHIUM_SOURCES
 
-    dilithium/p_dilithium3.c
-    dilithium/p_dilithium3_asn1.c
-    dilithium/sig_dilithium3.c
+    dilithium/pqdsa.c
+          dilithium/p_pqdsa.c
+          dilithium/p_pqdsa_asn1.c
+          dilithium/ml_dsa.c
   )
 endif()
 
@@ -774,7 +775,7 @@ if(BUILD_TESTING)
     ecdh_extra/ecdh_test.cc
     dh_extra/dh_test.cc
     digest_extra/digest_test.cc
-    dilithium/p_dilithium_test.cc
+          dilithium/p_pqdsa_test.cc
     dsa/dsa_test.cc
     des/des_test.cc
     endian_test.cc
