+++ crypto/evp_extra/internal.h
@@ -7,7 +7,7 @@
 #include <openssl/base.h>
 #include "../fipsmodule/evp/internal.h"
 
-#include "../dilithium/sig_dilithium.h"
+#include "../dilithium/ml_dsa.h"
 
 #define PKCS8_VERSION_ONE 0
 #define PKCS8_VERSION_TWO 1
